Read/write Objects <> Archive Files with Data Verification

Read and write objects to/from archival storage in the native RData format, with verification of data integrity.

archive(..., object, file)

Arguments

...
unnamed arguments will be parsed as 'file' if there is only one argument, 'object' and 'file' if there are two arguments.

object
(for write operations) an object.

file
the name of the file or a connection where the object is saved to or read from.

Value

For read operations: the object originally archived.

For write operations: invisible NULL. 'object' is written to 'file' as a side effect.

Details

For read operations: specify only 'file'. 'file' is read and the return value may be assigned to an object. A confirmation message is issued if the file read operation has been successful.

For write operations: specify both 'object' and 'file'. 'object' will be written to 'file'. A confirmation message is issued if the file write operation has been successful.

Data Verification

If the 'openssl' package is available, a sha256 hash of the original object is written to the archive. This allows the data integrity of the restored object to be verified when the archive is read back.

For write operations: confirmation of the sha256 hash is displayed if this has been successfully written to file.

For read operations: a 'data verified' message is issued if the sha256 hash found within the data file has been authenticated.

Examples

cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
file <- tempfile()

archive(cloud, file)

#> Archive written to '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpIZYel4/fileb6d4fbe0d58'
#> sha256: 3ba9b2c884032e9e264697bf980b15998df31dd993614b5b704af48ecb8d1bbe

restored <- archive(file)

#> Archive read from '/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpIZYel4/fileb6d4fbe0d58'
#> Data verified by sha256: 3ba9b2c884032e9e264697bf980b15998df31dd993614b5b704af48ecb8d1bbe

unlink(file)