Automated Ichimoku Strategies

Generate a list of the top performing ichimoku cloud strategies based on rule conditions of the form 'c1 > c2' or level 2 strategies based on combined rule conditions of the form 'c1 > c2 & c3 > c4'.

autostrat(x, n = 8, dir = c("long", "short"), level = 1)

Arguments

x
an ichimoku object.

n
defaults to 8. Select top n number of strategies to return.

dir
trade direction either 'long' or 'short' with a default of 'long'.

level
defaults to 1. Set to 2 to also return combined strategies.

Value

A list of 'n' ichimoku objects containing strategies. The cumulative log returns for all strategies as well as the summaries for the 'n' top strategies are saved as attributes to the list. The strategy summaries are printed to the console as a side effect.

Details

Ichimoku objects for each strategy are returned as a list. The cumulative log returns for all strategies as well as the summaries for the 'n' top strategies are saved as attributes to the list and may be viewed by using attributes() on the returned list.

Each individual ichimoku object may be accessed via its position in the list e.g. [[1]] for the 1st item.

Further Details

Please refer to the strategies vignette by running: vignette("strategies", package = "ichimoku")

Examples

cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

stratlist <- autostrat(cloud, n = 3)

#>                        [,1]               [,2]                [,3]             
#> Strategy               "senkouB > tenkan" "cloudBase > kijun" "senkouB > kijun"
#> ---------------------  "----------"       "----------"        "----------"     
#> Strategy cuml return % 14.72              11.96               11.7             
#> Per period mean ret %  0.0886             0.0729              0.0714           
#> Periods in market      57                 38                  57               
#> Total trades           3                  3                   3                
#> Average trade length   19                 12.67               19               
#> Trade success %        66.67              66.67               66.67            
#> Worst trade ret %      -0.31              -0.31               -0.31            
#> ---------------------  "----------"       "----------"        "----------"     
#> Benchmark cuml ret %   0.94               0.94                0.94             
#> Per period mean ret %  0.006              0.006               0.006            
#> Periods in market      155                155                 155              
#> ---------------------  "----------"       "----------"        "----------"     
#> Direction              "long"             "long"              "long"           
#> Start                  2020-04-02         2020-04-02          2020-04-02       
#> End                    2020-09-29         2020-09-29          2020-09-29       
#> Ticker                 "TKR"              "TKR"               "TKR"            

attributes(stratlist)

#> $logret
#>   senkouB_tenkan  cloudBase_kijun    senkouB_kijun       low_tenkan 
#>     0.0995100937     0.0991419543     0.0968956037     0.0887172560 
#> cloudBase_tenkan    senkouA_kijun        low_kijun    cloudBase_low 
#>     0.0873621713     0.0774384001     0.0769393981     0.0754852717 
#>  cloudBase_close   cloudTop_kijun    senkouB_close       high_kijun 
#>     0.0754366525     0.0751920495     0.0736018508     0.0648184687 
#>      senkouB_low      close_kijun      high_tenkan     senkouA_high 
#>     0.0623185299     0.0572281363     0.0525922102     0.0474448365 
#>    cloudTop_high   cloudBase_high     senkouB_high  cloudTop_tenkan 
#>     0.0471474786     0.0453936555     0.0450962976     0.0426371292 
#>     close_tenkan     kijun_chikou     chikou_close     tenkan_kijun 
#>     0.0418092675     0.0398283528     0.0394256746     0.0391080770 
#>   chikou_senkouA  chikou_cloudTop      chikou_high  senkouB_senkouA 
#>     0.0363964224     0.0363964224     0.0351623629     0.0337288705 
#>    tenkan_chikou      senkouA_low    senkouA_close   senkouA_tenkan 
#>     0.0336029689     0.0321489583     0.0321003390     0.0304892068 
#>   cloudTop_close   chikou_senkouB chikou_cloudBase     cloudTop_low 
#>     0.0302655373     0.0256804720     0.0256804720     0.0189822164 
#>       chikou_low     low_cloudTop   senkouB_chikou cloudBase_chikou 
#>     0.0171450323     0.0166369032     0.0107195926     0.0107195926 
#>       low_chikou   close_cloudTop   tenkan_senkouA    close_senkouA 
#>     0.0092856613     0.0053535823     0.0051299128     0.0035187806 
#>      low_senkouA     tenkan_close    chikou_tenkan  senkouA_senkouB 
#>     0.0034701613     0.0028189220     0.0020161507     0.0018902491 
#>   senkouA_chikou  cloudTop_chikou     close_chikou      high_chikou 
#>    -0.0007773028    -0.0007773028    -0.0038065550    -0.0039453047 
#>     chikou_kijun  tenkan_cloudTop      tenkan_high     high_senkouB 
#>    -0.0042092332    -0.0070180096    -0.0071505711    -0.0094771780 
#>   high_cloudBase    high_cloudTop     high_senkouA     kijun_tenkan 
#>    -0.0097745359    -0.0115283590    -0.0118257169    -0.0148570222 
#>      kijun_close      low_senkouB       kijun_high    close_senkouB 
#>    -0.0216090168    -0.0266994103    -0.0291993491    -0.0379827312 
#>   kijun_cloudTop  close_cloudBase    low_cloudBase        kijun_low 
#>    -0.0395729299    -0.0398175329    -0.0398661521    -0.0413202785 
#>    kijun_senkouA tenkan_cloudBase       tenkan_low    kijun_senkouB 
#>    -0.0418192806    -0.0517430517    -0.0530981364    -0.0542698381 
#>  kijun_cloudBase   tenkan_senkouB 
#>    -0.0565161887    -0.0638909741 
#> 
#> $summary
#>                        [,1]               [,2]                [,3]             
#> Strategy               "senkouB > tenkan" "cloudBase > kijun" "senkouB > kijun"
#> ---------------------  "----------"       "----------"        "----------"     
#> Strategy cuml return % 14.72              11.96               11.7             
#> Per period mean ret %  0.0886             0.0729              0.0714           
#> Periods in market      57                 38                  57               
#> Total trades           3                  3                   3                
#> Average trade length   19                 12.67               19               
#> Trade success %        66.67              66.67               66.67            
#> Worst trade ret %      -0.31              -0.31               -0.31            
#> ---------------------  "----------"       "----------"        "----------"     
#> Benchmark cuml ret %   0.94               0.94                0.94             
#> Per period mean ret %  0.006              0.006               0.006            
#> Periods in market      155                155                 155              
#> ---------------------  "----------"       "----------"        "----------"     
#> Direction              "long"             "long"              "long"           
#> Start                  2020-04-02         2020-04-02          2020-04-02       
#> End                    2020-09-29         2020-09-29          2020-09-29       
#> Ticker                 "TKR"              "TKR"               "TKR"            
#> 

summary(stratlist[[1]])

#>                        [,1]              
#> Strategy               "senkouB > tenkan"
#> ---------------------  "----------"      
#> Strategy cuml return % 14.72             
#> Per period mean ret %  0.0886            
#> Periods in market      57                
#> Total trades           3                 
#> Average trade length   19                
#> Trade success %        66.67             
#> Worst trade ret %      -0.31             
#> ---------------------  "----------"      
#> Benchmark cuml ret %   0.94              
#> Per period mean ret %  0.006             
#> Periods in market      155               
#> ---------------------  "----------"      
#> Direction              "long"            
#> Start                  2020-04-02        
#> End                    2020-09-29        
#> Ticker                 "TKR"             

autostrat(cloud, n = 1, dir = "short", level = 2)

#>                        [,1]                              
#> Strategy               "high > chikou & tenkan > senkouB"
#> ---------------------  "----------"                      
#> Strategy cuml return % 13.38                             
#> Per period mean ret %  0.0811                            
#> Periods in market      18                                
#> Total trades           3                                 
#> Average trade length   6                                 
#> Trade success %        0                                 
#> Worst trade ret %      -7.68                             
#> ---------------------  "----------"                      
#> Benchmark cuml ret %   -0.93                             
#> Per period mean ret %  -0.006                            
#> Periods in market      155                               
#> ---------------------  "----------"                      
#> Direction              "short"                           
#> Start                  2020-04-02                        
#> End                    2020-09-29                        
#> Ticker                 "TKR"