Plot Ichimoku Objects with ggplot2 and gtable

Plot Ichimoku Kinko Hyo cloud charts from ichimoku objects with a sub-plot for oscillators or a custom specified variable.

extraplot(
  object,
  window,
  ticker,
  subtitle,
  theme = c("original", "conceptual", "dark", "fresh", "mono", "solarized"),
  strat = TRUE,
  type = c("none", "r", "s", "bar", "line"),
  custom,
  ...
)

Arguments

object
an object of class 'ichimoku'.

window
(optional) a date-time window to subset the plot, in ISO-8601 compatible range strings of the format used for 'xts' objects, for example "2020-02-15/2020-08-15" or "2020-02-15/", "/2020-08" or "2020-07".

ticker
(optional) specify a ticker (or other text) to include in the chart heading. If not set, the ticker saved within the ichimoku object will be used.

subtitle
(optional) specify a subtitle to display under the chart title.

theme
[default 'original'] with alternative choices of 'conceptual', 'dark', 'fresh', 'mono', or 'solarized'.

strat
[default TRUE] if the ichimoku object contains a strategy, the periods for which the strategy results in a position will be shaded, and the strategy printed as the chart subtitle (if not otherwise specified). Set to FALSE to turn off this behaviour.

type
[default 'none'] type of sub-plot to display beneath the ichimoku cloud chart, with a choice of 'none', 'r' or 's' for the corresponding oscillator type, and 'bar' or 'line' for custom plots.

custom
(optional) character string (containing a regular expression) matching the column name of the variable to be displayed as sub-plot. Specify type = 'bar' or type = 'line', otherwise other type settings will take precedence.

...
other arguments not used by this method.

Value

A gtable object with S3 classes 'gtable', 'gTree', 'grob' and 'gDesc', or else a ggplot2 object with S3 classes 'gg' and 'ggplot' when falling back to a standard plot.

Details

The oscillator choices are between 'R-type', which is a modified form of a relative strength index (RSI), and 'S-type', which is a modified form of a stochastic oscillator. The oscillator look-back parameters are based on the fast and medium cloud periods of the ichimoku object.