Read ichimoku objects from Arrow Archive

Used to read ichimoku objects from archive storage in the Apache Arrow IPC file format.

ichimoku_read(file)

Arguments

file
A character file name or URI, raw vector, an Arrow input stream, or a FileSystem with path (SubTreeFileSystem). If a file name or URI, an Arrow InputStream will be opened and closed when finished. If an input stream is provided, it will be left open.

Value

The ichimoku object that was originally archived.

Details

Used to read ichimoku objects archived using the ichimoku_write function.

Examples

cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

filename <- tempfile()
ichimoku_write(cloud, filename)

#> ichimoku object 'cloud' written to: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpagW4AR/file166048dbaa25

restored <- ichimoku_read(filename)
all.equal(cloud, restored)

#> [1] "Attributes: < Component “index”: Attributes: < Component “tzone”: 1 string mismatch > >"

unlink(filename)