Look at Ichimoku Objects

Inspect the informational attributes of objects created by the 'ichimoku' package. Can also be used to extract ichimoku objects from lists returned by autostrat.

look(x, which)

Arguments

x
an ichimoku object or an object returned by autostrat, mlgrid or oanda.

which
(optional) integer value of strategy to return from an autostrat list.

Value

List of attribute values, or if 'which' is specified on an autostrat list, an ichimoku object containing a strategy.

Details

Note: for a level 2 autostrat object, if the object fails to print correctly due to its length, please access the list items directly using look(x)$summary and look(x)$logret, possibly in conjunction with head() or by setting the 'max' argument in print().

Examples

cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
look(cloud)

#> $periods
#> [1]  9 26 52
#> 
#> $periodicity
#> [1] 86400
#> 
#> $ticker
#> [1] "TKR"
#> 

stratlist <- autostrat(cloud, n = 3)

#>                        [,1]                [,2]               
#> Strategy               "senkouB > tenkan"  "cloudBase > kijun"
#> ---------------------  "----------"        "----------"       
#> Strategy cuml return % 14.72               11.96              
#> Per period mean ret %  0.0886              0.0729             
#> Periods in market      57                  38                 
#> Total trades           3                   3                  
#> Average trade length   19                  12.67              
#> Trade success %        66.67               66.67              
#> Worst trade ret %      -0.31               -0.31              
#> ---------------------  "----------"        "----------"       
#> Benchmark cuml ret %   0.94                0.94               
#> Per period mean ret %  0.006               0.006              
#> Periods in market      155                 155                
#> ---------------------  "----------"        "----------"       
#> Direction              "long"              "long"             
#> Start                  2020-04-01 23:00:00 2020-04-01 23:00:00
#> End                    2020-09-28 23:00:00 2020-09-28 23:00:00
#> Ticker                 "TKR"               "TKR"              
#>                        [,3]               
#> Strategy               "senkouB > kijun"  
#> ---------------------  "----------"       
#> Strategy cuml return % 11.7               
#> Per period mean ret %  0.0714             
#> Periods in market      57                 
#> Total trades           3                  
#> Average trade length   19                 
#> Trade success %        66.67              
#> Worst trade ret %      -0.31              
#> ---------------------  "----------"       
#> Benchmark cuml ret %   0.94               
#> Per period mean ret %  0.006              
#> Periods in market      155                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-01 23:00:00
#> End                    2020-09-28 23:00:00
#> Ticker                 "TKR"              

look(stratlist)

#> $logret
#>                         logret
#> senkouB_tenkan    0.0995100937
#> cloudBase_kijun   0.0991419543
#> senkouB_kijun     0.0968956037
#> low_tenkan        0.0887172560
#> cloudBase_tenkan  0.0873621713
#> senkouA_kijun     0.0774384001
#> low_kijun         0.0769393981
#> cloudBase_low     0.0754852717
#> cloudBase_close   0.0754366525
#> cloudTop_kijun    0.0751920495
#> senkouB_close     0.0736018508
#> high_kijun        0.0648184687
#> senkouB_low       0.0623185299
#> close_kijun       0.0572281363
#> high_tenkan       0.0525922102
#> senkouA_high      0.0474448365
#> cloudTop_high     0.0471474786
#> cloudBase_high    0.0453936555
#> senkouB_high      0.0450962976
#> cloudTop_tenkan   0.0426371292
#> close_tenkan      0.0418092675
#> kijun_chikou      0.0398283528
#> chikou_close      0.0394256746
#> tenkan_kijun      0.0391080770
#> chikou_senkouA    0.0363964224
#> chikou_cloudTop   0.0363964224
#> chikou_high       0.0351623629
#> senkouB_senkouA   0.0337288705
#> tenkan_chikou     0.0336029689
#> senkouA_low       0.0321489583
#> senkouA_close     0.0321003390
#> senkouA_tenkan    0.0304892068
#> cloudTop_close    0.0302655373
#> chikou_senkouB    0.0256804720
#> chikou_cloudBase  0.0256804720
#> cloudTop_low      0.0189822164
#> chikou_low        0.0171450323
#> low_cloudTop      0.0166369032
#> senkouB_chikou    0.0107195926
#> cloudBase_chikou  0.0107195926
#> low_chikou        0.0092856613
#> close_cloudTop    0.0053535823
#> tenkan_senkouA    0.0051299128
#> close_senkouA     0.0035187806
#> low_senkouA       0.0034701613
#> tenkan_close      0.0028189220
#> chikou_tenkan     0.0020161507
#> senkouA_senkouB   0.0018902491
#> senkouA_chikou   -0.0007773028
#> cloudTop_chikou  -0.0007773028
#> close_chikou     -0.0038065550
#> high_chikou      -0.0039453047
#> chikou_kijun     -0.0042092332
#> tenkan_cloudTop  -0.0070180096
#> tenkan_high      -0.0071505711
#> high_senkouB     -0.0094771780
#> high_cloudBase   -0.0097745359
#> high_cloudTop    -0.0115283590
#> high_senkouA     -0.0118257169
#> kijun_tenkan     -0.0148570222
#> kijun_close      -0.0216090168
#> low_senkouB      -0.0266994103
#> kijun_high       -0.0291993491
#> close_senkouB    -0.0379827312
#> kijun_cloudTop   -0.0395729299
#> close_cloudBase  -0.0398175329
#> low_cloudBase    -0.0398661521
#> kijun_low        -0.0413202785
#> kijun_senkouA    -0.0418192806
#> tenkan_cloudBase -0.0517430517
#> tenkan_low       -0.0530981364
#> kijun_senkouB    -0.0542698381
#> kijun_cloudBase  -0.0565161887
#> tenkan_senkouB   -0.0638909741
#> 
#> $summary
#>                        [,1]                [,2]               
#> Strategy               "senkouB > tenkan"  "cloudBase > kijun"
#> ---------------------  "----------"        "----------"       
#> Strategy cuml return % 14.72               11.96              
#> Per period mean ret %  0.0886              0.0729             
#> Periods in market      57                  38                 
#> Total trades           3                   3                  
#> Average trade length   19                  12.67              
#> Trade success %        66.67               66.67              
#> Worst trade ret %      -0.31               -0.31              
#> ---------------------  "----------"        "----------"       
#> Benchmark cuml ret %   0.94                0.94               
#> Per period mean ret %  0.006               0.006              
#> Periods in market      155                 155                
#> ---------------------  "----------"        "----------"       
#> Direction              "long"              "long"             
#> Start                  2020-04-01 23:00:00 2020-04-01 23:00:00
#> End                    2020-09-28 23:00:00 2020-09-28 23:00:00
#> Ticker                 "TKR"               "TKR"              
#>                        [,3]               
#> Strategy               "senkouB > kijun"  
#> ---------------------  "----------"       
#> Strategy cuml return % 11.7               
#> Per period mean ret %  0.0714             
#> Periods in market      57                 
#> Total trades           3                  
#> Average trade length   19                 
#> Trade success %        66.67              
#> Worst trade ret %      -0.31              
#> ---------------------  "----------"       
#> Benchmark cuml ret %   0.94               
#> Per period mean ret %  0.006              
#> Periods in market      155                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-01 23:00:00
#> End                    2020-09-28 23:00:00
#> Ticker                 "TKR"              
#> 

strat <- look(stratlist, which = 1)
look(strat)

#> $periods
#> [1]  9 26 52
#> 
#> $periodicity
#> [1] 86400
#> 
#> $ticker
#> [1] "TKR"
#> 
#> $strat
#>                        [,1]               
#> Strategy               "senkouB > tenkan" 
#> ---------------------  "----------"       
#> Strategy cuml return % 14.72              
#> Per period mean ret %  0.0886             
#> Periods in market      57                 
#> Total trades           3                  
#> Average trade length   19                 
#> Trade success %        66.67              
#> Worst trade ret %      -0.31              
#> ---------------------  "----------"       
#> Benchmark cuml ret %   0.94               
#> Per period mean ret %  0.006              
#> Periods in market      155                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-01 23:00:00
#> End                    2020-09-28 23:00:00
#> Ticker                 "TKR"              
#> 

grid <- mlgrid(cloud)
look(grid)

#> $y
#> [1] "logret"
#> 
#> $direction
#> [1] "long"
#> 
#> $ticker
#> [1] "TKR"
#> 

if (FALSE) {
# OANDA API key required to run this example
prices <- oanda("USD_JPY")
look(prices)
}