OANDA Studio Interactive Live Analysis

Interactive and fully-customisable R Shiny environment providing real-time Ichimoku Kinko Hyo cloud charts for major currencies, metals, commodities, government bonds and stock indices using OANDA fxTrade API data. Intuitive cursor infotip provides ready access to the data directly from the chart.

oanda_studio(
  instrument = "USD_JPY",
  granularity = c("D", "W", "M", "H12", "H8", "H6", "H4", "H3", "H2", "H1", "M30",
    "M15", "M10", "M5", "M4", "M2", "M1", "S30", "S15", "S10", "S5"),
  refresh = 5,
  count = 300,
  price = c("M", "B", "A"),
  theme = c("original", "dark", "solarized", "mono"),
  server = c("practice", "live"),
  apikey,
  ...,
  launch.browser = TRUE,
  periods = c(9L, 26L, 52L)
)

Arguments

instrument
[default 'USD_JPY'] string containing the base currency and quote currency delimited by a '_'. Use the oanda_instruments function to return a list of all valid instruments.

granularity
[default "D"] the granularity of the price data to fetch, one of "M", "W", "D", "H12", "H8", "H6", "H4", "H3", "H2", "H1", "M30", "M15", "M10", "M5", "M4", "M2", "M1", "S30", "S15", "S10", "S5".

refresh
[default 5] data refresh interval in seconds, with a minimum of 1.

count
[default 300] the number of periods to return, from 100 to 800. Note that fewer periods are actually shown on the chart to ensure a full cloud is always displayed.

price
[default "M"] pricing component, one of "M" (midpoint), "B" (bid) or "A" (ask).

theme
[default 'original'] chart theme with alternative choices of 'dark', 'solarized' or 'mono'.

server
[default "practice"] select either the "practice" or "live" server depending on the account type held with OANDA.

apikey
(optional) string containing the OANDA fxTrade API key (personal access token), or function that returns this string. Does not need to be specified if already stored by oanda_set_key(). Can also be entered interactively if not specified.

...
additional arguments passed along to ichimoku for calculating the ichimoku cloud, autoplot to set chart parameters, or the 'options' argument of shiny::shinyApp().

launch.browser
[default TRUE] If TRUE, the system's default web browser will be launched automatically after the app is started. The value of this argument can also be a function to call with the application's URL. To use the default Shiny viewer in RStudio, please specify getOption("shiny.launch.browser").

periods
[default c(9L, 26L, 52L)] a vector defining the length of periods used for the cloud. This parameter shoud not normally be modified as using other values would be invalid in the context of traditional ichimoku analysis.

Value

Returns a Shiny app object with class 'shiny.appobj'.

Details

This function polls the OANDA fxTrade API for the latest prices and updates a customisable reactive Shiny app at each refresh interval.

For further details please refer to the OANDA fxTrade API vignette by running: vignette("xoanda", package = "ichimoku").

Examples

if (FALSE) {
# OANDA fxTrade API key required to run this example
oanda_studio()

# To open in RStudio viewer instead of default browser
oanda_studio(launch.browser = getOption("shiny.launch.browser"))
}