Relative Numeric Representation

Produce a statistical summary of the latest numeric representation of the ichimoku cloud chart relative to historical values. Aids in determining if current trading falls within normal or exceptional ranges.

relative(x)

Arguments

x
an ichimoku object.

Value

A data frame containing a statistical summary of the latest ichimoku cloud chart representation and how it relates to the historical data. The elements are ordered by their absolute 'r value', so that those with the highest relative deviation are listed first. The time index value for 'latest' and the number of observations are printed to the console.

Details

Column 'mean' is the mean value, 'sd' the standard deviation, and 'latest' the latest observed values from the data.

The 'relative' or 'r value' is calculated as (latest - mean) / sd and represents a standardised (centred and scaled) measure of deviation.

Column 'p >= |r|' contains the empirical probability of the observed absolute 'r value' or greater.

Column 'E(|r|)|p' contains the expected absolute 'r value', conditional upon being greater than equal to the observed absolute 'r value'.

Further Details

Please refer to the strategies vignette by calling: vignette("strategies", package = "ichimoku")

Examples

cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
relative(cloud)

#> Latest: 2020-08-04 23:00:00 | n: 155 

#>                  mean   sd latest r value p >= |r| E(|r|)|p
#> chikou_high      0.79 6.09   6.60    0.95     0.36     1.46
#> low_tenkan      -0.30 1.72   1.30    0.93     0.40     1.38
#> low_kijun        0.44 3.28   3.45    0.92     0.43     1.37
#> chikou_close     1.51 6.07   7.00    0.91     0.39     1.41
#> chikou_low       2.31 6.12   7.80    0.90     0.41     1.37
#> low_cloudT       0.45 5.76   5.35    0.85     0.43     1.39
#> chikou_tenkan    1.73 6.15   6.90    0.84     0.41     1.41
#> close_kijun      1.27 3.25   3.85    0.79     0.48     1.31
#> close_cloudT     1.27 5.73   5.75    0.78     0.43     1.39
#> low_senkouA      1.18 5.97   5.50    0.72     0.45     1.36
#> high_cloudT      1.96 5.64   5.95    0.71     0.48     1.32
#> high_kijun       1.96 3.12   4.05    0.67     0.53     1.25
#> close_tenkan     0.52 1.77   1.70    0.67     0.52     1.24
#> close_senkouA    2.00 5.93   5.90    0.66     0.49     1.30
#> tenkan_cloudT    0.75 5.20   4.05    0.63     0.50     1.31
#> low_senkouB      2.19 5.38   5.35    0.59     0.64     1.16
#> high_senkouA     2.69 5.84   6.10    0.58     0.56     1.22
#> tenkan_kijun     0.75 2.42   2.15    0.58     0.60     1.16
#> kijun_cloudT     0.00 3.62   1.90    0.52     0.64     1.16
#> close_senkouB    3.02 5.34   5.75    0.51     0.73     1.09
#> tenkan_senkouA   1.48 5.44   4.20    0.50     0.57     1.21
#> low_cloudB       2.92 5.38   5.50    0.48     0.68     1.12
#> high_tenkan      1.21 1.56   1.90    0.44     0.66     1.06
#> chikou_kijun     2.28 5.96   4.90    0.44     0.63     1.15
#> high_senkouB     3.71 5.29   5.95    0.42     0.76     1.06
#> close_cloudB     3.75 5.31   5.90    0.40     0.70     1.11
#> senkouA_senkouB  1.01 2.98  -0.15   -0.39     0.70     1.10
#> kijun_senkouA    0.74 3.64   2.05    0.36     0.69     1.11
#> high_cloudB      4.44 5.27   6.10    0.32     0.71     1.10
#> tenkan_senkouB   2.49 5.12   4.05    0.30     0.79     1.03
#> chikou_cloudT    2.67 6.44   4.35    0.26     0.84     1.02
#> chikou_senkouA   3.44 6.46   4.75    0.20     0.89     0.97
#> tenkan_cloudB    3.23 5.12   4.20    0.19     0.79     1.01
#> kijun_cloudB     2.48 3.86   2.05   -0.11     0.98     0.86
#> chikou_cloudB    5.03 5.20   4.75   -0.05     0.99     0.89
#> kijun_senkouB    1.75 4.16   1.90    0.04     0.97     0.84
#> chikou_senkouB   4.26 5.40   4.35    0.02     1.00     0.88