Summary of Ichimoku Strategies

Custom summary method for ichimoku objects for viewing strategies.

# S3 method for ichimoku
summary(object, strat = TRUE, ...)

Arguments

object
an object of class 'ichimoku'.

strat
[default TRUE] to show the strategy summary if present. Set to FALSE to show the data summary instead.

...
additional arguments to be passed along.

Value

A matrix containing the strategy summary if present, otherwise a table containing the data summary.

Details

This function is an S3 method for the generic function summary() for class 'ichimoku'. It can be invoked by calling summary(x) on an object 'x' of class 'ichimoku'.

Further Details

Please refer to the strategies vignette by running: vignette("strategies", package = "ichimoku")

Examples

strat <- strat(ichimoku(sample_ohlc_data, ticker = "TKR"))
summary(strat)

#>                        [,1]               
#> Strategy               "close > tenkan"   
#> ---------------------  "----------"       
#> Strategy cuml return % 5.01               
#> Per period mean ret %  0.0218             
#> Periods in market      119                
#> Total trades           18                 
#> Average trade length   6.61               
#> Trade success %        27.78              
#> Worst trade ret %      -2.54              
#> ---------------------  "----------"       
#> Benchmark cuml ret %   3.38               
#> Per period mean ret %  0.0148             
#> Periods in market      224                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-01-13         
#> End                    2020-09-28 23:00:00
#> Ticker                 "TKR"