Select Trading Days

Used by ichimoku to subset a vector of dates to trading days.

tradingDays(x, holidays, ...)

Arguments

x
a vector of POSIXct date objects.

holidays
(optional) a vector, or function which outputs a vector, of dates defined as holidays.

...
other arguments not used by this function.

Value

A vector of logical values: TRUE if the corresponding element of 'x' is a weekday and not a holiday, FALSE otherwise.

Details

New Year's Day (01/01) and Christmas Day (25/12) are defined as holidays by default regardless of the values supplied by 'holidays'.

Examples

dates <- seq(from = as.POSIXct("2020-01-01"), by = "1 day", length.out = 7)
dates

#> [1] "2020-01-01 UTC" "2020-01-02 UTC" "2020-01-03 UTC" "2020-01-04 UTC"
#> [5] "2020-01-05 UTC" "2020-01-06 UTC" "2020-01-07 UTC"

tradingDays(dates)

#> [1] FALSE  TRUE  TRUE FALSE FALSE  TRUE  TRUE

tradingDays(dates, holidays = c("2020-01-02", "2020-01-03"))

#> [1] FALSE FALSE FALSE FALSE FALSE  TRUE  TRUE