Automated Ichimoku Strategies

Generate a list of the top performing ichimoku cloud strategies based on simple indicator conditions of the form 'c1 > c2' (level 1), complex combined strategies of the form 'c1 > c2 & c3 > c4' (level 2), or complex asymmetric strategies of the form 'c1 > c2 x c3 > c4' (level 3).

autostrat(x, n = 8, dir = c("long", "short"), level = 1)

Arguments

x
an ichimoku object.

n
[default 8] select top 'n' number of strategies to return.

dir
[default 'long'] trade direction, either 'long' or 'short'.

level
[default 1] to return simple strategies. For complex strategies, set level to 2 to return combined strategies of the form 's1 & s2' or level to 3 to return asymmetric strategies of the form 's1 x s2'

Value

A list of 'n' ichimoku objects containing strategies, with attributes 'logret', a matrix of cumulative log returns for all strategies, and 'summary', a matrix of summaries for the top 'n' strategies. The strategy summaries are printed to the console as a side effect.

Details

Ichimoku objects for each strategy are returned as a list. The cumulative log returns for all strategies as well as the summaries for the top 'n' strategies are saved as attributes to the list. This information may be retrieved by using look on the returned list.

Each individual ichimoku object may be accessed via its position in the list, e.g. [[1]] for the 1st item, or by using look specifying the parameter 'which'.

Further Details

Please refer to the strategies vignette by running: vignette("strategies", package = "ichimoku")

Examples

cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")

stratlist <- autostrat(cloud, n = 3)

#>                        [,1]                [,2]               
#> Strategy               "senkouB > tenkan"  "cloudB > tenkan"  
#> ---------------------  "----------"        "----------"       
#> Strategy cuml return % 17.49               16.08              
#> Per period mean ret %  0.0906              0.0838             
#> Periods in market      63                  51                 
#> Total trades           3                   3                  
#> Average trade length   21                  17                 
#> Trade success %        100                 100                
#> Worst trade ret %      3.64                3.16               
#> ---------------------  "----------"        "----------"       
#> Benchmark cuml ret %   5.53                5.53               
#> Per period mean ret %  0.0302              0.0302             
#> Periods in market      178                 178                
#> ---------------------  "----------"        "----------"       
#> Direction              "long"              "long"             
#> Start                  2020-04-19 23:00:00 2020-04-19 23:00:00
#> End                    2020-12-23          2020-12-23         
#> Ticker                 "TKR"               "TKR"              
#>                        [,3]               
#> Strategy               "senkouB > kijun"  
#> ---------------------  "----------"       
#> Strategy cuml return % 14.1               
#> Per period mean ret %  0.0741             
#> Periods in market      64                 
#> Total trades           3                  
#> Average trade length   21.33              
#> Trade success %        100                
#> Worst trade ret %      3.49               
#> ---------------------  "----------"       
#> Benchmark cuml ret %   5.53               
#> Per period mean ret %  0.0302             
#> Periods in market      178                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-19 23:00:00
#> End                    2020-12-23         
#> Ticker                 "TKR"              

look(stratlist)

#> $logret
#>                       logret
#> senkouB_tenkan   0.144869048
#> cloudB_tenkan    0.132721126
#> senkouB_kijun    0.124552623
#> cloudB_kijun     0.124476092
#> low_tenkan       0.119988595
#> cloudB_close     0.117048881
#> high_tenkan      0.112062056
#> cloudB_low       0.110900734
#> senkouB_close    0.110699399
#> cloudT_kijun     0.103153955
#> senkouA_kijun    0.103077424
#> senkouB_low      0.101480718
#> cloudT_tenkan    0.093961798
#> cloudT_high      0.088759707
#> low_kijun        0.088503030
#> senkouB_high     0.086708526
#> chikou_senkouA   0.084344426
#> high_kijun       0.083373239
#> senkouA_tenkan   0.081813876
#> close_tenkan     0.079948752
#> tenkan_chikou    0.078961923
#> senkouA_high     0.076693196
#> cloudB_high      0.074642015
#> senkouA_low      0.073761187
#> senkouA_close    0.073712567
#> close_kijun      0.071380845
#> chikou_cloudT    0.069192621
#> cloudT_close     0.067363085
#> kijun_chikou     0.066288776
#> cloudT_low       0.064341171
#> chikou_cloudB    0.064291073
#> chikou_close     0.062225816
#> kijun_tenkan     0.055091789
#> chikou_low       0.053536566
#> chikou_senkouB   0.049139267
#> senkouA_senkouB  0.048655509
#> senkouB_chikou   0.045949698
#> chikou_high      0.042730532
#> low_chikou       0.041552399
#> high_chikou      0.034326207
#> close_chikou     0.032863149
#> cloudB_chikou    0.030797893
#> low_cloudT       0.030747795
#> tenkan_kijun     0.030128362
#> chikou_kijun     0.028800189
#> cloudT_chikou    0.025896344
#> close_cloudT     0.024713830
#> kijun_close      0.023708120
#> senkouB_senkouA  0.023618778
#> tenkan_close     0.023402736
#> close_senkouA    0.021376398
#> low_senkouA      0.021327779
#> high_cloudB      0.020446951
#> high_senkouA     0.018395770
#> tenkan_senkouA   0.013275089
#> senkouA_chikou   0.010744539
#> chikou_tenkan    0.010161327
#> high_senkouB     0.008380439
#> kijun_low        0.006585935
#> high_cloudT      0.006329258
#> tenkan_cloudT    0.001127167
#> kijun_high      -0.003436078
#> low_senkouB     -0.006391752
#> tenkan_high     -0.007150571
#> kijun_senkouA   -0.007988459
#> kijun_cloudT    -0.008064990
#> low_cloudB      -0.015811768
#> tenkan_low      -0.015890560
#> close_senkouB   -0.018622484
#> close_cloudB    -0.021959916
#> tenkan_cloudB   -0.037632160
#> kijun_cloudB    -0.040858512
#> kijun_senkouB   -0.040935043
#> tenkan_senkouB  -0.049780083
#> 
#> $summary
#>                        [,1]                [,2]               
#> Strategy               "senkouB > tenkan"  "cloudB > tenkan"  
#> ---------------------  "----------"        "----------"       
#> Strategy cuml return % 17.49               16.08              
#> Per period mean ret %  0.0906              0.0838             
#> Periods in market      63                  51                 
#> Total trades           3                   3                  
#> Average trade length   21                  17                 
#> Trade success %        100                 100                
#> Worst trade ret %      3.64                3.16               
#> ---------------------  "----------"        "----------"       
#> Benchmark cuml ret %   5.53                5.53               
#> Per period mean ret %  0.0302              0.0302             
#> Periods in market      178                 178                
#> ---------------------  "----------"        "----------"       
#> Direction              "long"              "long"             
#> Start                  2020-04-19 23:00:00 2020-04-19 23:00:00
#> End                    2020-12-23          2020-12-23         
#> Ticker                 "TKR"               "TKR"              
#>                        [,3]               
#> Strategy               "senkouB > kijun"  
#> ---------------------  "----------"       
#> Strategy cuml return % 14.1               
#> Per period mean ret %  0.0741             
#> Periods in market      64                 
#> Total trades           3                  
#> Average trade length   21.33              
#> Trade success %        100                
#> Worst trade ret %      3.49               
#> ---------------------  "----------"       
#> Benchmark cuml ret %   5.53               
#> Per period mean ret %  0.0302             
#> Periods in market      178                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-19 23:00:00
#> End                    2020-12-23         
#> Ticker                 "TKR"              
#> 

strat <- look(stratlist, which = 1)
summary(strat)

#>                        [,1]               
#> Strategy               "senkouB > tenkan" 
#> ---------------------  "----------"       
#> Strategy cuml return % 17.49              
#> Per period mean ret %  0.0906             
#> Periods in market      63                 
#> Total trades           3                  
#> Average trade length   21                 
#> Trade success %        100                
#> Worst trade ret %      3.64               
#> ---------------------  "----------"       
#> Benchmark cuml ret %   5.53               
#> Per period mean ret %  0.0302             
#> Periods in market      178                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-19 23:00:00
#> End                    2020-12-23         
#> Ticker                 "TKR"              

autostrat(cloud, n = 1, dir = "short", level = 2)

#>                        [,1]                               
#> Strategy               "close > chikou & tenkan > senkouB"
#> ---------------------  "----------"                       
#> Strategy cuml return % 11.21                              
#> Per period mean ret %  0.0597                             
#> Periods in market      20                                 
#> Total trades           4                                  
#> Average trade length   5                                  
#> Trade success %        75                                 
#> Worst trade ret %      -0.44                              
#> ---------------------  "----------"                       
#> Benchmark cuml ret %   -5.24                              
#> Per period mean ret %  -0.0302                            
#> Periods in market      178                                
#> ---------------------  "----------"                       
#> Direction              "short"                            
#> Start                  2020-04-19 23:00:00                
#> End                    2020-12-23                         
#> Ticker                 "TKR"                              

autostrat(cloud, n = 1, dir = "long", level = 3)

#>                        [,1]                             
#> Strategy               "senkouB > senkouA x kijun > low"
#> ---------------------  "----------"                     
#> Strategy cuml return % 2.49                             
#> Per period mean ret %  0.0138                           
#> Periods in market      59                               
#> Total trades           3                                
#> Average trade length   19.67                            
#> Trade success %        66.67                            
#> Worst trade ret %      -2.49                            
#> ---------------------  "----------"                     
#> Benchmark cuml ret %   5.53                             
#> Per period mean ret %  0.0302                           
#> Periods in market      178                              
#> ---------------------  "----------"                     
#> Direction              "long"                           
#> Start                  2020-04-19 23:00:00              
#> End                    2020-12-23                       
#> Ticker                 "TKR"