ichimoku: Visualization and Tools for Ichimoku Kinko Hyo Strategies

An implementation of 'Ichimoku Kinko Hyo', also commonly known as 'cloud charts'. Static and interactive visualizations with tools for creating, backtesting and development of quantitative 'ichimoku' strategies. As described in Sasaki (1996, ISBN:4925152009), the technique is a refinement on candlestick charting originating from Japan, now in widespread use in technical analysis worldwide. Translating as 'one-glance equilibrium chart', it allows the price action and market structure of financial securities to be determined 'at-a-glance'. Incorporates an interface with the OANDA fxTrade API https://developer.oanda.com/ for retrieving historical and live streaming price data for major currencies, metals, commodities, government bonds and stock indices.

Principal ichimoku functions

Data

ichimoku to create an ichimoku object from price data.

archive for reading and writing ichimoku objects to/from Apache Arrow archive files.

oanda to retrieve price data from the OANDA fxTrade API.

oanda_stream to stream a live data feed from the OANDA fxTrade API.

Visualization

plot.ichimoku to plot a cloud chart from an ichimoku object.

iplot to plot an interactive cloud chart from an ichimoku object.

oanda_chart to create live updating ichimoku cloud charts using OANDA data.

oanda_studio a complete live analysis environment using OANDA data implemented in R Shiny.

Strategies & ML

strat to augment an ichimoku object with a strategy, including complex combined and asymmetric strategies.

stratcombine to create custom combined strategies.

autostrat to automatically evaluate and rank top-performing strategies.

mlgrid to generate a numeric representation of the relationship between ichimoku cloud chart elements.

References

Sasaki, H. 佐々木 英信 (1996), 一目均衡表の研究 [ichimoku kinkouhyou no kenkyuu]. Tokyo, Japan: Toushi Radar.

OANDA' and 'fxTrade' are trademarks owned by OANDA Corporation, an entity unaffiliated with the ichimoku package.

Gao, C. (2021), ichimoku: Visualization and Tools for Ichimoku Kinko Hyo Strategies. R package version 0.3.51. https://shikokuchuo.net/ichimoku/ CRAN page: https://CRAN.R-project.org/package=ichimoku

Author

Charlie Gao <charlie.gao@shikokuchuo.net>