ichimoku: Visualization and Tools for Ichimoku Kinko Hyo Strategies

An implementation of 'Ichimoku Kinko Hyo', also commonly known as 'cloud charts'. Static and interactive visualizations with tools for creating, backtesting and development of quantitative 'ichimoku' strategies. As described in Sasaki (1996, ISBN:4925152009), the technique is a refinement on candlestick charting, originating from Japan and now in widespread use in technical analysis worldwide. Translating as 'one-glance equilibrium chart', it allows the price action and market structure of financial securities to be determined 'at-a-glance'. Incorporates an interface with the OANDA fxTrade API https://developer.oanda.com/ for retrieving historical and live streaming price data for major currencies, metals, commodities, government bonds and stock indices.

Principal ichimoku functions

Data & Visualization

ichimoku to create an ichimoku object from price data.

plot.ichimoku / iplot to plot (interactive) cloud charts from ichimoku objects.

archive for reading/writing objects to/from archive files with data verification.

oanda to retrieve price data from the OANDA fxTrade API.

Strategies & ML

strat to augment an ichimoku object with a strategy, including combined and asymmetric strategies.

autostrat to automatically evaluate and rank top-performing strategies.

mlgrid to generate a numeric representation of the ichimoku cloud chart.

relative to produce a statistical summary of the latest ichimoku numeric representation relative to historical values.

Real-time

oanda_chart to plot real-time ichimoku cloud charts using OANDA data.

oanda_studio a complete live analysis environment using OANDA data implemented in R Shiny.

oanda_stream / oanda_quote to obtain the latest live data stream / quote from the OANDA fxTrade API.

oanda_view for a market overview showing the relative performance of constituents.

oanda_orders / oanda_positions to retrieve the aggregate OANDA fxTrade order / position book.

References

Sasaki, H. (1996), ichimoku kinkouhyou no kenkyuu. Tokyo, Japan: Toushi Radar.

OANDA' and 'fxTrade' are trademarks owned by OANDA Corporation, an entity unaffiliated with the ichimoku package, its authors or copyright holders.

See also

Author

Maintainer: Charlie Gao charlie.gao@shikokuchuo.net (ORCID)

Other contributors:

Hibiki AI Limited [copyright holder]