Look at Informational Attributes

Inspect the informational attributes of objects, or extract ichimoku objects from lists returned by autostrat.

look(x, which)

Arguments

x
an object (created by the ichimoku package).

which
(optional) integer value of strategy to return from an autostrat list.

Value

For objects created by the ichimoku package, a list of attribute values specific to that data type, or if 'which' is specified on an autostrat list, an ichimoku object containing a strategy.

For other objects, a list of attributes that are non-standard for 'matrix', 'data.frame', or 'xts' objects, or else invisible NULL if none are present.

Details

Note: for a level 2 autostrat object, if the object fails to print correctly due to its length, please access the list items directly using look(x)$summary and look(x)$logret, possibly in conjunction with head() or by setting the 'max' argument in print().

Examples

cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
look(cloud)

#> $periods
#> [1]  9 26 52
#> 
#> $periodicity
#> [1] 86400
#> 
#> $ticker
#> [1] "TKR"
#> 

stratlist <- autostrat(cloud, n = 3)

#>                        [,1]                [,2]               
#> Strategy               "senkouB > tenkan"  "cloudB > tenkan"  
#> ---------------------  "----------"        "----------"       
#> Strategy cuml return % 17.49               16.08              
#> Per period mean ret %  0.0906              0.0838             
#> Periods in market      63                  51                 
#> Total trades           3                   3                  
#> Average trade length   21                  17                 
#> Trade success %        100                 100                
#> Worst trade ret %      3.64                3.16               
#> ---------------------  "----------"        "----------"       
#> Benchmark cuml ret %   5.53                5.53               
#> Per period mean ret %  0.0302              0.0302             
#> Periods in market      178                 178                
#> ---------------------  "----------"        "----------"       
#> Direction              "long"              "long"             
#> Start                  2020-04-19 23:00:00 2020-04-19 23:00:00
#> End                    2020-12-23          2020-12-23         
#> Ticker                 "TKR"               "TKR"              
#>                        [,3]               
#> Strategy               "senkouB > kijun"  
#> ---------------------  "----------"       
#> Strategy cuml return % 14.1               
#> Per period mean ret %  0.0741             
#> Periods in market      64                 
#> Total trades           3                  
#> Average trade length   21.33              
#> Trade success %        100                
#> Worst trade ret %      3.49               
#> ---------------------  "----------"       
#> Benchmark cuml ret %   5.53               
#> Per period mean ret %  0.0302             
#> Periods in market      178                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-19 23:00:00
#> End                    2020-12-23         
#> Ticker                 "TKR"              

look(stratlist)

#> $logret
#>                       logret
#> senkouB_tenkan   0.144869048
#> cloudB_tenkan    0.132721126
#> senkouB_kijun    0.124552623
#> cloudB_kijun     0.124476092
#> low_tenkan       0.119988595
#> cloudB_close     0.117048881
#> high_tenkan      0.112062056
#> cloudB_low       0.110900734
#> senkouB_close    0.110699399
#> cloudT_kijun     0.103153955
#> senkouA_kijun    0.103077424
#> senkouB_low      0.101480718
#> cloudT_tenkan    0.093961798
#> cloudT_high      0.088759707
#> low_kijun        0.088503030
#> senkouB_high     0.086708526
#> chikou_senkouA   0.084344426
#> high_kijun       0.083373239
#> senkouA_tenkan   0.081813876
#> close_tenkan     0.079948752
#> tenkan_chikou    0.078961923
#> senkouA_high     0.076693196
#> cloudB_high      0.074642015
#> senkouA_low      0.073761187
#> senkouA_close    0.073712567
#> close_kijun      0.071380845
#> chikou_cloudT    0.069192621
#> cloudT_close     0.067363085
#> kijun_chikou     0.066288776
#> cloudT_low       0.064341171
#> chikou_cloudB    0.064291073
#> chikou_close     0.062225816
#> kijun_tenkan     0.055091789
#> chikou_low       0.053536566
#> chikou_senkouB   0.049139267
#> senkouA_senkouB  0.048655509
#> senkouB_chikou   0.045949698
#> chikou_high      0.042730532
#> low_chikou       0.041552399
#> high_chikou      0.034326207
#> close_chikou     0.032863149
#> cloudB_chikou    0.030797893
#> low_cloudT       0.030747795
#> tenkan_kijun     0.030128362
#> chikou_kijun     0.028800189
#> cloudT_chikou    0.025896344
#> close_cloudT     0.024713830
#> kijun_close      0.023708120
#> senkouB_senkouA  0.023618778
#> tenkan_close     0.023402736
#> close_senkouA    0.021376398
#> low_senkouA      0.021327779
#> high_cloudB      0.020446951
#> high_senkouA     0.018395770
#> tenkan_senkouA   0.013275089
#> senkouA_chikou   0.010744539
#> chikou_tenkan    0.010161327
#> high_senkouB     0.008380439
#> kijun_low        0.006585935
#> high_cloudT      0.006329258
#> tenkan_cloudT    0.001127167
#> kijun_high      -0.003436078
#> low_senkouB     -0.006391752
#> tenkan_high     -0.007150571
#> kijun_senkouA   -0.007988459
#> kijun_cloudT    -0.008064990
#> low_cloudB      -0.015811768
#> tenkan_low      -0.015890560
#> close_senkouB   -0.018622484
#> close_cloudB    -0.021959916
#> tenkan_cloudB   -0.037632160
#> kijun_cloudB    -0.040858512
#> kijun_senkouB   -0.040935043
#> tenkan_senkouB  -0.049780083
#> 
#> $summary
#>                        [,1]                [,2]               
#> Strategy               "senkouB > tenkan"  "cloudB > tenkan"  
#> ---------------------  "----------"        "----------"       
#> Strategy cuml return % 17.49               16.08              
#> Per period mean ret %  0.0906              0.0838             
#> Periods in market      63                  51                 
#> Total trades           3                   3                  
#> Average trade length   21                  17                 
#> Trade success %        100                 100                
#> Worst trade ret %      3.64                3.16               
#> ---------------------  "----------"        "----------"       
#> Benchmark cuml ret %   5.53                5.53               
#> Per period mean ret %  0.0302              0.0302             
#> Periods in market      178                 178                
#> ---------------------  "----------"        "----------"       
#> Direction              "long"              "long"             
#> Start                  2020-04-19 23:00:00 2020-04-19 23:00:00
#> End                    2020-12-23          2020-12-23         
#> Ticker                 "TKR"               "TKR"              
#>                        [,3]               
#> Strategy               "senkouB > kijun"  
#> ---------------------  "----------"       
#> Strategy cuml return % 14.1               
#> Per period mean ret %  0.0741             
#> Periods in market      64                 
#> Total trades           3                  
#> Average trade length   21.33              
#> Trade success %        100                
#> Worst trade ret %      3.49               
#> ---------------------  "----------"       
#> Benchmark cuml ret %   5.53               
#> Per period mean ret %  0.0302             
#> Periods in market      178                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-19 23:00:00
#> End                    2020-12-23         
#> Ticker                 "TKR"              
#> 

strat <- look(stratlist, which = 1)
look(strat)

#> $periods
#> [1]  9 26 52
#> 
#> $periodicity
#> [1] 86400
#> 
#> $ticker
#> [1] "TKR"
#> 
#> $strat
#>                        [,1]               
#> Strategy               "senkouB > tenkan" 
#> ---------------------  "----------"       
#> Strategy cuml return % 17.49              
#> Per period mean ret %  0.0906             
#> Periods in market      63                 
#> Total trades           3                  
#> Average trade length   21                 
#> Trade success %        100                
#> Worst trade ret %      3.64               
#> ---------------------  "----------"       
#> Benchmark cuml ret %   5.53               
#> Per period mean ret %  0.0302             
#> Periods in market      178                
#> ---------------------  "----------"       
#> Direction              "long"             
#> Start                  2020-04-19 23:00:00
#> End                    2020-12-23         
#> Ticker                 "TKR"              
#> 

grid <- mlgrid(cloud)
look(grid)

#> $y
#> [1] "logret"
#> 
#> $direction
#> [1] "long"
#> 
#> $ticker
#> [1] "TKR"
#> 

if (FALSE) {
# OANDA API key required to run this example
prices <- oanda("USD_JPY")
look(prices)
}