mlgrid Numeric Representation

Create a grid of ichimoku rule conditions and next period returns. The grid facilitates comparing strategy returns or as a basis for further processing in machine learning applications. The purpose of this function is to translate the visual representation of the relationship between cloud chart elements into a numerical format for further analysis.

mlgrid(
  x,
  y = c("logret", "ret"),
  dir = c("long", "short"),
  type = c("boolean", "numeric"),
  unique = TRUE
)

Arguments

x
an ichimoku object.

y
choose target variable 'logret' (log returns) or 'ret' (discrete returns), with a default of 'logret'.

dir
trade direction either 'long' or 'short' with a default of 'long'.

type
either 'boolean' or 'numeric' with a default of 'boolean'. 'boolean' creates a grid of dummy variables for ichimoku indicator conditions of the form 1 if c1 > c2, 0 otherwise. 'numeric' creates a grid of the numeric difference c1 - c2.

unique
defaults to TRUE to return only unique combinations of c1 and c2. Set to FALSE to return both c1 > c2 and c2 > c1.

Value

A data.frame in a 'tidy' format with one observation per row and one feature per column with the target 'y' as the first column.

The 'y' parameter and trade direction are set as atrributes. To view these, use the attributes() function on the returned object.

Details

The date-time index corresponds to when the condition is met at the close for that period. The return is the single-period return achieved by transacting at the immediately following opening price until the next opening price.

Only valid combinations are included. This excludes any combination involving 'open' as it is in effect a lagged indicator and not contemporaneous. The following trivial or highly-collinear pairs are also excluded: (high, close) ,(low, close), (low, high), (cloudTop, Senkou A), (cloudBase, senkou A), (cloudTop, senkouB), (cloudBase, senkouB), (cloudBase, cloudTop).

mlgrid is used by autostrat to enumerate the returns for all valid strategy combinations.

Further Details

Please refer to the strategies vignette by running: vignette("strategies", package = "ichimoku")

Examples

cloud <- ichimoku(sample_ohlc_data, ticker = "TKR")
grid <- mlgrid(cloud, y = "ret", dir = "short", type = "numeric")
attributes(grid)

#> $names
#>  [1] "y"                "chikou_close"     "chikou_high"      "chikou_low"      
#>  [5] "chikou_tenkan"    "chikou_kijun"     "chikou_senkouA"   "chikou_senkouB"  
#>  [9] "chikou_cloudTop"  "chikou_cloudBase" "close_tenkan"     "close_kijun"     
#> [13] "close_senkouA"    "close_senkouB"    "close_cloudTop"   "close_cloudBase" 
#> [17] "high_tenkan"      "high_kijun"       "high_senkouA"     "high_senkouB"    
#> [21] "high_cloudTop"    "high_cloudBase"   "low_tenkan"       "low_kijun"       
#> [25] "low_senkouA"      "low_senkouB"      "low_cloudTop"     "low_cloudBase"   
#> [29] "tenkan_kijun"     "tenkan_senkouA"   "tenkan_senkouB"   "tenkan_cloudTop" 
#> [33] "tenkan_cloudBase" "kijun_senkouA"    "kijun_senkouB"    "kijun_cloudTop"  
#> [37] "kijun_cloudBase"  "senkouA_senkouB" 
#> 
#> $row.names
#>   [1] "2020-05-01" "2020-05-03" "2020-05-04" "2020-05-05" "2020-05-06"
#>   [6] "2020-05-07" "2020-05-08" "2020-05-10" "2020-05-11" "2020-05-12"
#>  [11] "2020-05-13" "2020-05-14" "2020-05-15" "2020-05-17" "2020-05-18"
#>  [16] "2020-05-19" "2020-05-20" "2020-05-21" "2020-05-22" "2020-05-24"
#>  [21] "2020-05-25" "2020-05-26" "2020-05-27" "2020-05-28" "2020-05-29"
#>  [26] "2020-05-31" "2020-06-01" "2020-06-02" "2020-06-03" "2020-06-04"
#>  [31] "2020-06-05" "2020-06-07" "2020-06-08" "2020-06-09" "2020-06-10"
#>  [36] "2020-06-11" "2020-06-12" "2020-06-14" "2020-06-15" "2020-06-16"
#>  [41] "2020-06-17" "2020-06-18" "2020-06-19" "2020-06-21" "2020-06-22"
#>  [46] "2020-06-23" "2020-06-24" "2020-06-25" "2020-06-26" "2020-06-28"
#>  [51] "2020-06-29" "2020-06-30" "2020-07-01" "2020-07-02" "2020-07-03"
#>  [56] "2020-07-05" "2020-07-06" "2020-07-07" "2020-07-08" "2020-07-09"
#>  [61] "2020-07-10" "2020-07-12" "2020-07-13" "2020-07-14" "2020-07-15"
#>  [66] "2020-07-16" "2020-07-17" "2020-07-19" "2020-07-20" "2020-07-21"
#>  [71] "2020-07-22" "2020-07-23" "2020-07-24" "2020-07-26" "2020-07-27"
#>  [76] "2020-07-28" "2020-07-29" "2020-07-30" "2020-07-31" "2020-08-02"
#>  [81] "2020-08-03" "2020-08-04" "2020-08-05" "2020-08-06" "2020-08-07"
#>  [86] "2020-08-09" "2020-08-10" "2020-08-11" "2020-08-12" "2020-08-13"
#>  [91] "2020-08-14" "2020-08-16" "2020-08-17" "2020-08-18" "2020-08-19"
#>  [96] "2020-08-20" "2020-08-21" "2020-08-23" "2020-08-24" "2020-08-25"
#> [101] "2020-08-26" "2020-08-27" "2020-08-28" "2020-08-30" "2020-08-31"
#> [106] "2020-09-01" "2020-09-02" "2020-09-03" "2020-09-04" "2020-09-06"
#> [111] "2020-09-07" "2020-09-08" "2020-09-09" "2020-09-10" "2020-09-11"
#> [116] "2020-09-13" "2020-09-14" "2020-09-15" "2020-09-16" "2020-09-17"
#> [121] "2020-09-18" "2020-09-20" "2020-09-21" "2020-09-22" "2020-09-23"
#> [126] "2020-09-24" "2020-09-25" "2020-09-27" "2020-09-28"
#> 
#> $class
#> [1] "data.frame"
#> 
#> $y
#> [1] "ret"
#> 
#> $direction
#> [1] "short"
#> 
#> $mlgrid
#> [1] TRUE
#>