OANDA Streaming Data

Stream live price and liquidity data for major currencies, metals, commodities, government bonds and stock indices from the OANDA fxTrade Streaming API.

oanda_stream(instrument, server = c("practice", "live"), apikey)

Arguments

instrument
string containing the base currency and quote currency delimited by a '_' (for example "USD_JPY"). Use the oanda_instruments function to return a list of all valid instruments.

server
[default "practice"] select either the "practice" or "live" server depending on the account type held with OANDA.

apikey
(optional) string containing the OANDA fxTrade API key (personal access token), or function that returns this string. Does not need to be specified if already stored by oanda_set_key(). Can also be entered interactively if not specified.

Value

Does not return a value, however the streaming data is output to the console as a side effect.

Details

This function connects to the OANDA fxTrade Streaming API. Use the 'Esc' key to stop the stream.

For further details please refer to the OANDA fxTrade API vignette by running: vignette("xoanda", package = "ichimoku").

Streaming Data

Get a stream of Account Prices starting from when the request is made. This pricing stream does not include every single price created for the Account, but instead will provide at most 4 prices per second (every 250 milliseconds) for each instrument being requested. If more than one price is created for an instrument during the 250 millisecond window, only the price in effect at the end of the window is sent. This means that during periods of rapid price movement, subscribers to this stream will not be sent every price. Pricing windows for different connections to the price stream are not all aligned in the same way (i.e. they are not all aligned to the top of the second). This means that during periods of rapid price movement, different subscribers may observe different prices depending on their alignment.

The streamed data is formatted for easier reading, but otherwise represents the raw feed without omission. Note that as this is a raw stream, returned times are in UTC.

Examples

if (FALSE) {
# OANDA fxTrade API key required to run this example
oanda_stream("USD_JPY")
}